ceph: only raise conditions that represent current state
The conditions on the cephcluster CR were being set to false
when not in progress, which is misleading to the purpose of
conditions. The conditions will now always represent current
state of the cluster. Transient conditions that show progress of
a reconcile will be removed from the status.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 15d0537cec54b503aea21beb0084d0e6dd3464eb)

# Conflicts:
#	pkg/operator/ceph/config/conditions.go