ceph:  placement in case of both PVC and non-PVC's
In the case of PVC,
We are giving lower priority to all placement.
We want deviceSet placement to applied and
override in case of overlapping settings and
we are merging nodeAffinity if applied in both
all placement and deviceSet.

In case of non-PVC,
we apply spec.placement

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 319e4a41a411ba0a890f3aaee49f4ed8b05ffb69)