ceph: allow null values for pool settings
The schema in v1.6 is more strict and we are seeing errors in the csv
upgrades where nullable values are not allowed for the pool settings.
We now allow the nullable settings again to fix the upgrade scenario.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit f22d8d4dc94ce0f73599aa84493017a058dd1caa)

# Conflicts:
#	cluster/examples/kubernetes/ceph/crds.yaml