docs: clarify not to change crdsenabled in helm chart
Changing the crds.enabled setting in a running cluster will cause the
CRs to be deleted, which could be disastrous for a cluster. Add
and clarify the warning in the docs to prevent it and link
to recovery steps if it does happen.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 5e86ef8b0c7e6f683f84bde0cf179d5686f2eadb)