ceph: run CLI output with Info logging
In e8f9cfcb714685d93017be279d9ffdfc3847acca, the logging was replaced by
Debug which is probably a mistake given the intent of the commit to
enable debug logging on the prepare job.
However, this code is only triggered when running a ceph-osd with the
rook binary, so we must run Info and run Debug since Debug is not
activated.

Signed-off-by: Sébastien Han <seb@redhat.com>
(cherry picked from commit da56354aa3c18ff3b88848aaea9c1849c79f18d5)