ceph: update cluster-on-pvc example for proper OSD scheduling
The topology spread constraint example allowed scheduling even
if the TSC could not be satisfied when scheduling the osd prepare
jobs. This could lead to unevenly spread OSDs especially when
the cluster was not using the zone labels. Now the topology will be
required for scheduling, to ensure proper spread of the OSDs.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>